move node-replacing to PostRector
